requirements.txt
iniconfig==2.1.0
numpy==2.2.4
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
python-flint==0.7.1
z3-solver==4.14.1.0
